+++ root_utils/root_file_utils.py
@@ -121,9 +121,9 @@ class WCSim:
                 pmt.append(pmt_id)
                 trigger.append(t)
         hits = {
-            "position": np.asarray(position, dtype=np.float32),
-            "charge": np.asarray(charge, dtype=np.float32),
-            "time": np.asarray(time, dtype=np.float32),
+            "position": np.asarray(position, dtype=np.float64),
+            "charge": np.asarray(charge, dtype=np.float64),
+            "time": np.asarray(time, dtype=np.float64),
             "pmt": np.asarray(pmt, dtype=np.int32),
             "trigger": np.asarray(trigger, dtype=np.int32)
         }
@@ -149,7 +149,7 @@ class WCSim:
                 PE.append(hit.GetTotalPe(1))
                 trigger.append(t)
         hits = {
-            "position": np.asarray(position, dtype=np.float32),
+            "position": np.asarray(position, dtype=np.float64),
             "track": np.asarray(track, dtype=np.int32),
             "pmt": np.asarray(pmt, dtype=np.int32),
             "PE": np.asarray(PE, dtype=np.int32),
@@ -184,10 +184,10 @@ class WCSim:
                     pmt.append(pmt_id)
                     trigger.append(t)
         photons = {
-            "start_position": np.asarray(start_position, dtype=np.float32),
-            "end_position": np.asarray(end_position, dtype=np.float32),
-            "start_time": np.asarray(start_time, dtype=np.float32),
-            "end_time": np.asarray(end_time, dtype=np.float32),
+            "start_position": np.asarray(start_position, dtype=np.float64),
+            "end_position": np.asarray(end_position, dtype=np.float64),
+            "start_time": np.asarray(start_time, dtype=np.float64),
+            "end_time": np.asarray(end_time, dtype=np.float64),
             "track": np.asarray(track, dtype=np.int32),
             "pmt": np.asarray(pmt, dtype=np.int32),
             "trigger": np.asarray(trigger, dtype=np.int32)
@@ -217,17 +217,17 @@ class WCSim:
         tracks = {
             "id": np.asarray(id, dtype=np.int32),
             "pid": np.asarray(pid, dtype=np.int32),
-            "start_time": np.asarray(start_time, dtype=np.float32),
-            "energy": np.asarray(energy, dtype=np.float32),
-            "start_position": np.asarray(start_position, dtype=np.float32),
-            "stop_position": np.asarray(stop_position, dtype=np.float32),
+            "start_time": np.asarray(start_time, dtype=np.float64),
+            "energy": np.asarray(energy, dtype=np.float64),
+            "start_position": np.asarray(start_position, dtype=np.float64),
+            "stop_position": np.asarray(stop_position, dtype=np.float64),
             "parent": np.asarray(parent, dtype=np.int32),
             "flag": np.asarray(flag, dtype=np.int32)
         }
         return tracks
 
     def get_trigger_times(self):
-        trigger_times = np.empty(self.ntrigger, dtype=np.float32)
+        trigger_times = np.empty(self.ntrigger, dtype=np.float64)
         for t in range(self.ntrigger):
             self.get_trigger(t)
             trigger_times[t] = self.trigger.GetHeader().GetDate()
